Extract file observation related methods to separate files.
